[#1360] Move fetch monitoring initialValue to Forms.jsx from MonitoringDetail.jsx
